requirements.txt
numpy>=2.0
matplotlib>=3.8
pillow>=10.0
gradio>=5.0
